refactor(terrain): split Terrain implementation into responsibility modules
terrain.cpp was 3064 lines. Move method bodies (no API/header change) into terrain_io.cpp (file loaders + QSC metadata), terrain_lod.cpp (update + render-cube selection), terrain_mesh.cpp (cube data/mesh generation), and terrain_query.cpp (height queries + editor raycast). Shared helpers and math constants move to a private terrain_internal.h. Core terrain.cpp keeps lifecycle, static tables, and pool accessors (246 lines). New heavy TUs added to the PCH-skip list. Pure code motion: builds clean and loads terrain at runtime (level 2, 558 objects, no errors).
